examples: remove dot assignment support
This is an opiniated change, yet at this point it seems warrented.

The dot syntax adds a separate way of using graphviz.
It requires various complexities to be supported
It would be possible to add a similar behavior using current syntax (not done here)
We want to have views which will support such behavior